- Changed: Additional memory reductions for game livings and some changes to the previous commit.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk/GameServer@2412 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ gameutils/INpcTemplate.cs
@@ -164,12 +164,12 @@ namespace DOL.GS
 		/// </summary>
 		int AggroRange { get;}
 
-		int Race { get; }
+		ushort Race { get; }
 
 		/// <summary>
 		/// Gets the body type of the npc
 		/// </summary>
-		int BodyType { get;}
+		ushort BodyType { get;}
 		
 		/// <summary>
 		/// The Mob's max distance from its spawn before return automatically
